pyproject.toml
[tool.poetry]
name = "excel generate"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "2.2.1"
fastapi = "0.110.0"
python-dotenv = "1.0.1"
pydantic-settings = "2.2.1"
uvicorn = "0.28.0"
python-multipart = "0.0.9"
openpyxl = "3.1.2"
lxml = "5.1.0"
aiohttp = "3.9.5"
annotated-types = "0.6.0"
anyio = "4.3.0"
build = "1.1.1"
cachecontrol = "0.14.0"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
click = "8.1.7"
colorama = "0.4.6"
crashtest = "0.4.1"
distlib = "0.3.8"
dulwich = "0.21.7"
et-xmlfile = "1.1.0"
fastjsonschema = "2.19.1"
filelock = "3.13.1"
h11 = "0.14.0"
idna = "3.6"
installer = "0.7.0"
jaraco-classes = "3.3.1"
keyring = "24.3.1"
more-itertools = "10.2.0"
msgpack = "1.0.8"
numpy = "1.26.4"
packaging = "23.2"
pexpect = "4.9.0"
pkginfo = "1.10.0"
platformdirs = "4.2.0"
poetry = "1.8.2"
poetry-core = "1.9.0"
poetry-plugin-export = "1.6.0"
ptyprocess = "0.7.0"
pydantic = "2.6.4"
pydantic-core = "2.16.3"
pyproject-hooks = "1.0.0"
python-dateutil = "2.9.0.post0"
pytz = "2024.1"
pywin32-ctypes = "0.2.2"
rapidfuzz = "3.6.2"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
six = "1.16.0"
sniffio = "1.3.1"
starlette = "0.36.3"
tomlkit = "0.12.4"
trove-classifiers = "2024.3.3"
typing-extensions = "4.10.0"
tzdata = "2024.1"
urllib3 = "2.2.1"
virtualenv = "20.25.1"
aiosignal = "1.3.1"
attrs = "23.2.0"
black = "24.4.2"
frozenlist = "1.4.1"
isort = "5.13.2"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
pathspec = "0.12.1"
pillow = "10.3.0"
yarl = "1.9.4"
python-docx = "1.1.2"
python-pptx = "0.6.23"
xlsxwriter = "3.2.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
